测试推送服务功能

+++ apps/fileupload/views.py
@@ -49,18 +49,26 @@ class  UploadView(LoginRequiredMixin,View):
                 from django.core.urlresolvers import reverse
                 return HttpResponseRedirect(reverse('salt:salt_deploy_dev'))
             elif os.path.exists(SAVE_DIR) == True:
-                f = open(os.path.join(SAVE_DIR, filesource.name), 'wb')
-                for line in filesource.chunks():
-                    f.write(line)
-                f.close()
-                all_files = os.walk(SAVE_DIR)
-                UploadFiles.objects.all().delete()
-                for dirname,dirs,file in all_files:
-                    for filename in file:
-                        file_db = UploadFiles.objects.create(name=str(filename))
-                        file_db.save()
-                from django.core.urlresolvers import reverse
-                return HttpResponseRedirect(reverse('salt:salt_deploy_dev'))
+                if str(filesource).split('.')[-1] == "gz" or str(filesource).split('.')[-1] == "zip":
+                    f = open(os.path.join(SAVE_DIR, filesource.name), 'wb')
+                    for line in filesource.chunks():
+                        f.write(line)
+                    f.close()
+                    all_files = os.walk(SAVE_DIR)
+                    UploadFiles.objects.all().delete()
+                    for dirname,dirs,file in all_files:
+                        for filename in file:
+                            file_db = UploadFiles.objects.create(name=str(filename))
+                            file_db.save()
+                    from django.core.urlresolvers import reverse
+                    return HttpResponseRedirect(reverse('salt:salt_deploy_dev'))
+                else:
+                    msg_file = u'格式必须为tar.gz或者zip！'
+                    return render(request, 'salt_deploy_dev.html', {
+                        'msg_file': msg_file,
+                        'all_service_dev': all_service_dev,
+                        'all_files': all_files,
+                    })
         else:
             from django.core.urlresolvers import reverse
             return HttpResponseRedirect(reverse('salt:salt_deploy_dev'))
@@ -79,34 +87,34 @@ class UploadGitView(LoginRequiredMixin,View):
             SAVE_DIR = os.path.join(MEDIA_ROOT, 'upload/sls')
             GIT_DIR = SAVE_DIR + '/' + git_file_dir
             if os.path.exists(GIT_DIR) == True:
-                rm_fr_git_dir_cmd = 'rm -fr ' + SAVE_DIR + '/' + git_file_dir
+                rm_fr_git_dir_cmd = 'rm -fr ' + GIT_DIR
                 rm_fr_git_dir = subprocess.Popen(rm_fr_git_dir_cmd, stdout=subprocess.PIPE, shell=True)
                 rm_fr_git_dir_stdout = rm_fr_git_dir.communicate()[0].strip()
-                GIT_TAR = SAVE_DIR + '/' + git_file_dir + '.tar.gz'
+                GIT_TAR = GIT_DIR + '.tar.gz'
                 if os.path.exists(GIT_TAR) == True:
                     rm_fr_git_tar_cmd = 'rm -fr ' + GIT_TAR
                     rm_fr_git_tar = subprocess.Popen(rm_fr_git_tar_cmd, stdout=subprocess.PIPE, shell=True)
                     rm_fr_git_tar_stdout = rm_fr_git_tar.communicate()[0].strip()
-                    re_get_git_file_cmd = 'git clone ' + str(git) + ' ' + SAVE_DIR
+                    re_get_git_file_cmd = 'git clone ' + str(git) + ' ' + '\"' +  GIT_DIR + '\"'
                     re_get_git_file = subprocess.Popen(re_get_git_file_cmd, stdout=subprocess.PIPE, shell=True)
                     re_get_git_file_stdout = re_get_git_file.communicate()[0].strip()
-                    tar_git_dir_cmd = 'tar zcf ' + SAVE_DIR + '/' + git_file_dir + '.tar.gz ' + SAVE_DIR + '/' + git_file_dir
+                    tar_git_dir_cmd = 'tar zcf ' + SAVE_DIR + '/' + git_file_dir + '.tar.gz -C ' + SAVE_DIR + " " + git_file_dir
                     tar_git_dir = subprocess.Popen(tar_git_dir_cmd, stdout=subprocess.PIPE, shell=True)
                     tar_git_dir_stdout = tar_git_dir.communicate()[0].strip()
-                    rm_fr_git_dir_cmd = 'rm -fr ' + SAVE_DIR + '/' + git_file_dir + '/'
+                    rm_fr_git_dir_cmd = 'rm -fr ' + GIT_DIR
                     rm_fr_git_dir = subprocess.Popen(rm_fr_git_dir_cmd, stdout=subprocess.PIPE, shell=True)
                     rm_fr_git_dir_stdout = rm_fr_git_dir.communicate()[0].strip()
                     git_file_db = UploadFiles.objects.get_or_create(name=str(git_file_dir))
                     from django.core.urlresolvers import reverse
                     return HttpResponseRedirect(reverse('salt:salt_deploy_dev'))
                 else:
-                    re_get_git_file_cmd = 'git clone ' + str(git) + ' ' + SAVE_DIR
+                    re_get_git_file_cmd = 'git clone ' + str(git) + ' ' + '\"' + GIT_DIR + '\"'
                     re_get_git_file = subprocess.Popen(re_get_git_file_cmd, stdout=subprocess.PIPE, shell=True)
                     re_get_git_file_stdout = re_get_git_file.communicate()[0].strip()
-                    tar_git_dir_cmd = 'tar zcf ' + SAVE_DIR + '/' + git_file_dir + '.tar.gz ' + SAVE_DIR + '/' + git_file_dir
+                    tar_git_dir_cmd = 'tar zcf ' + SAVE_DIR + '/' + git_file_dir + '.tar.gz -C ' + SAVE_DIR + " " + git_file_dir
                     tar_git_dir = subprocess.Popen(tar_git_dir_cmd, stdout=subprocess.PIPE, shell=True)
                     tar_git_dir_stdout = tar_git_dir.communicate()[0].strip()
-                    rm_fr_git_dir_cmd = 'rm -fr ' + SAVE_DIR + '/' + git_file_dir + '/'
+                    rm_fr_git_dir_cmd = 'rm -fr ' + GIT_DIR
                     rm_fr_git_dir = subprocess.Popen(rm_fr_git_dir_cmd, stdout=subprocess.PIPE, shell=True)
                     rm_fr_git_dir_stdout = rm_fr_git_dir.communicate()[0].strip()
                     git_file_db = UploadFiles.objects.get_or_create(name=str(git_file_dir))
@@ -118,29 +126,30 @@ class UploadGitView(LoginRequiredMixin,View):
                     rm_fr_git_tar_cmd = 'rm -fr ' + GIT_TAR
                     rm_fr_git_tar = subprocess.Popen(rm_fr_git_tar_cmd, stdout=subprocess.PIPE, shell=True)
                     rm_fr_git_tar_stdout = rm_fr_git_tar.communicate()[0].strip()
-                    re_get_git_file_cmd = 'git clone ' + str(git) + ' ' + SAVE_DIR
+                    re_get_git_file_cmd = 'git clone ' + str(git) + ' ' + '\"' + GIT_DIR + '\"'
+                    print re_get_git_file_cmd
                     re_get_git_file = subprocess.Popen(re_get_git_file_cmd, stdout=subprocess.PIPE, shell=True)
                     re_get_git_file_stdout = re_get_git_file.communicate()[0].strip()
-                    tar_git_dir_cmd = 'tar zcf ' + SAVE_DIR + '/' + git_file_dir + '.tar.gz ' + SAVE_DIR + '/' + git_file_dir
+                    tar_git_dir_cmd = 'tar zcf ' + SAVE_DIR + '/' + git_file_dir + '.tar.gz -C ' + SAVE_DIR + " " + git_file_dir
                     tar_git_dir = subprocess.Popen(tar_git_dir_cmd, stdout=subprocess.PIPE, shell=True)
                     tar_git_dir_stdout = tar_git_dir.communicate()[0].strip()
-                    rm_fr_git_dir_cmd = 'rm -fr ' + SAVE_DIR + '/' + git_file_dir + '/'
+                    rm_fr_git_dir_cmd = 'rm -fr ' + GIT_DIR
                     rm_fr_git_dir = subprocess.Popen(rm_fr_git_dir_cmd, stdout=subprocess.PIPE, shell=True)
                     rm_fr_git_dir_stdout = rm_fr_git_dir.communicate()[0].strip()
-                    git_file_db = UploadFiles.objects.get_or_create(name=str(git_file_dir))
+                    git_file_db = UploadFiles.objects.get_or_create(name=str(git_file_dir + '.tar.gz'))
                     from django.core.urlresolvers import reverse
                     return HttpResponseRedirect(reverse('salt:salt_deploy_dev'))
                 else:
-                    get_git_file_cmd = 'git clone ' + str(git) + ' ' + SAVE_DIR
+                    get_git_file_cmd = 'git clone ' + str(git) + ' ' + '\"' + GIT_DIR + '\"'
                     get_git_file = subprocess.Popen(get_git_file_cmd,stdout=subprocess.PIPE,shell=True)
                     get_git_file_stdout = get_git_file.communicate()[0].strip()
-                    tar_git_dir_cmd = 'tar zcf ' + SAVE_DIR + '/' + git_file_dir + '.tar.gz ' + SAVE_DIR + '/' + git_file_dir
+                    tar_git_dir_cmd = 'tar zcf ' + SAVE_DIR + '/' + git_file_dir + '.tar.gz -C ' + SAVE_DIR + " " + git_file_dir
                     tar_git_dir = subprocess.Popen(tar_git_dir_cmd,stdout=subprocess.PIPE,shell=True)
                     tar_git_dir_stdout = tar_git_dir.communicate()[0].strip()
-                    rm_fr_git_dir_cmd = 'rm -fr ' + SAVE_DIR + '/' + git_file_dir + '/'
+                    rm_fr_git_dir_cmd = 'rm -fr ' + GIT_DIR
                     rm_fr_git_dir = subprocess.Popen(rm_fr_git_dir_cmd,stdout=subprocess.PIPE,shell=True)
                     rm_fr_git_dir_stdout = rm_fr_git_dir.communicate()[0].strip()
-                    git_file_db = UploadFiles.objects.get_or_create(name=str(git_file_dir))
+                    git_file_db = UploadFiles.objects.get_or_create(name=str(git_file_dir + '.tar.gz'))
                     from django.core.urlresolvers import reverse
                     return HttpResponseRedirect(reverse('salt:salt_deploy_dev'))
 
@@ -188,6 +197,10 @@ class UploadRemoveView(LoginRequiredMixin,View):
         remove_file_cmd = "rm -fr " + str(file)
         remove_file = subprocess.Popen(remove_file_cmd,stdout=subprocess.PIPE,shell=True)
         remove_file_out = remove_file.communicate()[0].strip()
+        #UploadFiles.objects.filter(name=str(file_remove)).delete()
+        file_query = UploadFiles.objects.get(name=str(file_remove))
+        file_id = file_query.id
+        Service.objects.filter(file_id=file_id).update(file_id=None)
         UploadFiles.objects.filter(name=str(file_remove)).delete()
         from django.core.urlresolvers import reverse
         return HttpResponseRedirect(reverse('salt:salt_deploy_dev'))
